fix: updated owner id to bigInt

+++ services/restaurant/db_setup.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS restaurants (
     closing_time TIME NOT NULL,
     media_url VARCHAR(255),
     max_slots INTEGER NOT NULL CHECK (max_slots >= 0),
-    owner_id INTEGER NOT NULL,
+    owner_id BIGINT NOT NULL,
     owner_name VARCHAR(100) NOT NULL,
     sponsor_tier INTEGER NOT NULL DEFAULT 0 CHECK (sponsor_tier >= 0)
 );
@@ -22,4 +22,3 @@ CREATE TABLE IF NOT EXISTS restaurant_categories (
 
 CREATE INDEX IF NOT EXISTS idx_restaurant_categories_restaurant_id
     ON restaurant_categories (restaurant_id);
-
